backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.33.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.3.0
python-multipart==0.0.20
pytz==2024.1
slowapi==0.1.9
httpx==0.28.1
twilio==9.3.0
cryptography==42.0.0
sendgrid==6.11.0
